Add initial code for room detector project
Most of the code included is modified from the code in liblab, with
the notable exception of the USART interfaces, including low-level
functions, a byte queue receiver task, a byte queue transmitter
task, a packet encoder/decoder, and a packet queue receiver task.
Also included are two test programs for the master and compute
microcontrollers.  The compute microcontroller polls readings from
three analog sensors and transmits them using the packet encoding.
The master microcontroller receives and decodes packets and displays
their contained data on the LCD.